feat(version): move the version builtin to lib/version; identify the embedder
The (version) builtin leaves core for a new lib/version + nsversion
namespace and gains :main {:name :version} — the running program
itself, taken from Go build info (the embedder's main module) or from
the new version.SetMain(name, ver) API for -ldflags -X branding.
--version prints that identity first when the main module is not
jig/lisp, so an embedder's REPL built on command.Execute reports
itself; jig/lisp and jig/scanner stay listed as components.

core.Versions moved to lib/version unchanged; embedders keep the
builtin by loading nsversion (migration note in the CHANGELOG).
