+++ cmd/lisp/main.go
@@ -21,6 +21,7 @@ import (
 	"github.com/jig/lisp/lib/system/nssystem"
 	"github.com/jig/lisp/lib/term/nsterm"
 	"github.com/jig/lisp/lib/test/nstest"
+	"github.com/jig/lisp/lib/version/nsversion"
 	"github.com/jig/lisp/lib/web/nsweb"
 	"github.com/jig/lisp/types"
 )
@@ -58,6 +59,7 @@ func libraries(scriptArgs []string) []library {
 
 		// new libraries on jig/lisp v0.6.0
 		{"log", nslog.Load},
+		{"version", nsversion.Load},
 	}
 }
